Integrate network kernel
