+++ include/pops/kernel.hpp
@@ -75,10 +75,10 @@ namespace pops {
  * See NaturalAnthropogenicDispersalKernel and SwitchDispersalKernel for further
  * documentation.
  */
-template<typename IntegerRaster>
+template<typename IntegerRaster, typename RasterIndex>
 using DispersalKernel = NaturalAnthropogenicDispersalKernel<
-    SwitchDispersalKernel<IntegerRaster>,
-    SwitchDispersalKernel<IntegerRaster>>;
+    SwitchDispersalKernel<IntegerRaster, RasterIndex>,
+    SwitchDispersalKernel<IntegerRaster, RasterIndex>>;
 
 }  // namespace pops
