+++ include/pops/model.hpp
@@ -57,8 +57,8 @@ protected:
      * @param dispersers The disperser raster (reference, for deterministic kernel)
      * @return Created kernel
      */
-    SwitchDispersalKernel<IntegerRaster>
-    create_natural_kernel(const IntegerRaster& dispersers)
+    SwitchDispersalKernel<IntegerRaster, RasterIndex> create_natural_kernel(
+        const IntegerRaster& dispersers, const Network<RasterIndex>& network)
     {
         RadialDispersalKernel<IntegerRaster> radial_kernel(
             config_.ew_res,
@@ -76,11 +76,14 @@ protected:
             config_.ns_res,
             config_.natural_scale,
             config_.shape);
-        SwitchDispersalKernel<IntegerRaster> selectable_kernel(
+        NetworkDispersalKernel<RasterIndex> network_kernel(
+            network, config_.network_min_time, config_.network_max_time);
+        SwitchDispersalKernel<IntegerRaster, RasterIndex> selectable_kernel(
             natural_kernel,
             radial_kernel,
             deterministic_kernel,
             uniform_kernel,
+            network_kernel,
             natural_neighbor_kernel,
             config_.deterministic);
         return selectable_kernel;
@@ -96,8 +99,9 @@ protected:
      * @param dispersers The disperser raster (reference, for deterministic kernel)
      * @return Created kernel
      */
-    SwitchDispersalKernel<IntegerRaster>
-    create_overpopulation_movement_kernel(const IntegerRaster& dispersers)
+    SwitchDispersalKernel<IntegerRaster, RasterIndex>
+    create_overpopulation_movement_kernel(
+        const IntegerRaster& dispersers, const Network<RasterIndex>& network)
     {
         RadialDispersalKernel<IntegerRaster> radial_kernel(
             config_.ew_res,
@@ -115,11 +119,14 @@ protected:
             config_.ns_res,
             config_.natural_scale * config_.leaving_scale_coefficient,
             config_.shape);
-        SwitchDispersalKernel<IntegerRaster> selectable_kernel(
+        NetworkDispersalKernel<RasterIndex> network_kernel(
+            network, config_.network_min_time, config_.network_max_time);
+        SwitchDispersalKernel<IntegerRaster, RasterIndex> selectable_kernel(
             natural_kernel,
             radial_kernel,
             deterministic_kernel,
             uniform_kernel,
+            network_kernel,
             natural_neighbor_kernel,
             config_.deterministic);
         return selectable_kernel;
@@ -134,8 +141,8 @@ protected:
      * @param dispersers The disperser raster (reference, for deterministic kernel)
      * @return Created kernel
      */
-    SwitchDispersalKernel<IntegerRaster>
-    create_anthro_kernel(const IntegerRaster& dispersers)
+    SwitchDispersalKernel<IntegerRaster, RasterIndex> create_anthro_kernel(
+        const IntegerRaster& dispersers, const Network<RasterIndex>& network)
     {
         RadialDispersalKernel<IntegerRaster> radial_kernel(
             config_.ew_res,
@@ -153,11 +160,14 @@ protected:
             config_.ns_res,
             config_.anthro_scale,
             config_.shape);
-        SwitchDispersalKernel<IntegerRaster> selectable_kernel(
+        NetworkDispersalKernel<RasterIndex> network_kernel(
+            network, config_.network_min_time, config_.network_max_time);
+        SwitchDispersalKernel<IntegerRaster, RasterIndex> selectable_kernel(
             anthro_kernel,
             radial_kernel,
             deterministic_kernel,
             uniform_kernel,
+            network_kernel,
             anthro_neighbor_kernel,
             config_.deterministic);
         return selectable_kernel;
@@ -251,6 +261,7 @@ public:
         QuarantineEscape<IntegerRaster>& quarantine,  // out
         const IntegerRaster& quarantine_areas,
         const std::vector<std::vector<int>> movements,
+        const Network<RasterIndex>& network,
         std::vector<std::vector<int>>& suitable_cells)
     {
 
@@ -275,13 +286,13 @@ public:
                 config_.reproductive_rate,
                 suitable_cells);
 
-            DispersalKernel<IntegerRaster> dispersal_kernel(
-                create_natural_kernel(dispersers),
-                create_anthro_kernel(dispersers),
+            DispersalKernel<IntegerRaster, RasterIndex> dispersal_kernel(
+                create_natural_kernel(dispersers, network),
+                create_anthro_kernel(dispersers, network),
                 config_.use_anthropogenic_kernel,
                 config_.percent_natural_dispersal);
             auto overpopulation_kernel =
-                create_overpopulation_movement_kernel(dispersers);
+                create_overpopulation_movement_kernel(dispersers, network);
 
             simulation_.disperse_and_infect(
                 step,
