+++ include/pops/switch_kernel.hpp
@@ -20,6 +20,7 @@
 #include "deterministic_kernel.hpp"
 #include "uniform_kernel.hpp"
 #include "neighbor_kernel.hpp"
+#include "network_kernel.hpp"
 #include "kernel_types.hpp"
 #include "utils.hpp"
 
@@ -34,7 +35,7 @@ namespace pops {
  * its call in the function call operator, and extend the
  * supports_kernel() function.
  */
-template<typename IntegerRaster>
+template<typename IntegerRaster, typename RasterIndex>
 class SwitchDispersalKernel
 {
 protected:
@@ -43,6 +44,7 @@ protected:
     DeterministicDispersalKernel<IntegerRaster> deterministic_kernel_;
     UniformDispersalKernel uniform_kernel_;
     DeterministicNeighborDispersalKernel deterministic_neighbor_kernel_;
+    NetworkDispersalKernel<RasterIndex> network_kernel_;
     bool deterministic_;
 
 public:
@@ -51,6 +53,7 @@ public:
         const RadialDispersalKernel<IntegerRaster>& radial_kernel,
         const DeterministicDispersalKernel<IntegerRaster>& deterministic_kernel,
         const UniformDispersalKernel& uniform_kernel,
+        const NetworkDispersalKernel<RasterIndex>& network_kernel,
         const DeterministicNeighborDispersalKernel& deterministic_neighbor_kernel =
             DeterministicNeighborDispersalKernel(Direction::None),
         const bool deterministic = false)
@@ -61,6 +64,7 @@ public:
           deterministic_kernel_(deterministic_kernel),
           uniform_kernel_(uniform_kernel),
           deterministic_neighbor_kernel_(deterministic_neighbor_kernel),
+          network_kernel_(network_kernel),
           deterministic_(deterministic)
     {}
 
@@ -76,6 +80,9 @@ public:
         else if (dispersal_kernel_type_ == DispersalKernelType::DeterministicNeighbor) {
             return deterministic_neighbor_kernel_(generator, row, col);
         }
+        else if (dispersal_kernel_type_ == DispersalKernelType::Network) {
+            return network_kernel_(generator, row, col);
+        }
         else if (deterministic_) {
             return deterministic_kernel_(generator, row, col);
         }
@@ -86,10 +93,23 @@ public:
 
     bool is_cell_eligible(int row, int col)
     {
-        UNUSED(row);
-        UNUSED(col);
-        // TODO: delegate this to individual sub-kernels
-        return true;
+        // switch in between the supported kernels
+        if (dispersal_kernel_type_ == DispersalKernelType::Uniform) {
+            // TODO: Individual kernels should support this.
+            return true;
+        }
+        else if (dispersal_kernel_type_ == DispersalKernelType::DeterministicNeighbor) {
+            return true;
+        }
+        else if (dispersal_kernel_type_ == DispersalKernelType::Network) {
+            return network_kernel_.is_cell_eligible(row, col);
+        }
+        else if (deterministic_) {
+            return true;
+        }
+        else {
+            return true;
+        }
     }
 
     /*! \copydoc RadialDispersalKernel::supports_kernel()
