+++ tests/CMakeLists.txt
@@ -31,4 +31,5 @@ add_pops_test(test_statistics)
 add_pops_test(test_treatments)
 add_pops_test(test_overpopulation_movements)
 add_pops_test(test_quarantine)
+add_pops_test(test_network)
 add_pops_test(test_network_kernel)
